+++ examples/mainline.ts
@@ -26,7 +26,7 @@ export interface CliOption {
     help?: string;
 }
 
-var optionsNoAuth: CliOption[] = [
+const optionsNoAuth: CliOption[] = [
     {
         names: ['help', 'h'],
         type: 'bool',
@@ -49,7 +49,7 @@ var optionsNoAuth: CliOption[] = [
     }
 ];
 
-var options: CliOption[] = [
+const options: CliOption[] = [
     ...optionsNoAuth,
     {
         names: ['username', 'u'],
@@ -67,7 +67,7 @@ var options: CliOption[] = [
 export function fail(cmd: string, err: Error, opts: {
     verbose?: boolean;
 } = {}) {
-    var errToShow = opts.verbose ? err.stack || err : err.message || err;
+    const errToShow = opts.verbose ? err.stack || err : err.message || err;
     console.error('%s: error: %s', cmd, errToShow);
     Deno.exit(2);
 }
@@ -78,7 +78,7 @@ export function mainline(config: {
     excludeAuth?: boolean;
     options?: CliOption[];
 }) {
-    var dashOpts = (config.excludeAuth ? optionsNoAuth : options);
+    let dashOpts = (config.excludeAuth ? optionsNoAuth : options);
     if (config.options) {
         // Add to existing options.
         dashOpts = dashOpts.concat(config.options);